단위 테스트 & 메서드 수정
