+++ src/main/java/racing_car/Calculator.java
@@ -1,7 +1,23 @@
 package racing_car;
 
+import java.util.Arrays;
+
 public class Calculator {
     int result;
+    private String operator[] = {"+","-","*","/"};
+    public int validationCheck(String str[]){
+        int first =Integer.parseInt(str[0]);
+        result = first;
+        for (int i=1; i<str.length; i+=2){
+            String op =str[i];
+            if (!(Arrays.asList(operator).contains(op))){
+                throw new IllegalArgumentException("사칙 연산 기호가 아닙니다");
+            }
+            int second = Integer.parseInt(str[i + 1]);
+            result = operation(result, op , second);
+        }
+        return result;
+    }
     public int operation(int first , String operator , int second){
         if (operator.equals("+")) result = add(first,second);
         if (operator.equals("-")) result =subtract(first,second);
